add bagfix
add formated date from backend, change styled for Tytles from plate QA
