+++ src/components/UserData/UserPhoto/UserPhoto.styled.jsx
@@ -3,7 +3,7 @@ import { theme } from 'globalStyles/theme';
 import { ReactComponent as KameraIcon } from 'images/svg/kamera.svg';
 import { ReactComponent as CheckIcon } from 'images/svg/check.svg';
 
-export const AvatarWrapper = styled.div`
+export const AvatarWrapper = styled.form`
   @media (min-width: ${theme.breakpoints[2]}) {
     width: 382px;
     margin-left: auto;
